Refactor to ask password on reboot
It is necessary to always ask for the password even if the device is
unlocked, such as when rebooting from a previously unlocked session.

Without this the reboot will succeed without asking for the password,
but upon sleep/resume the OS will crash due to the disk password not
being stored in kernel memory.